# packages/contract/openapi.yaml
components:
  schemas:
...
    github_com_bingoring_forin_server_internal_domain_colleague.Relation:
      enum:
      - peer
      - mentor
      - mentee
      type: string
      x-enum-varnames:
      - RelationPeer
      - RelationMentor
      - RelationMentee
    github_com_bingoring_forin_server_internal_domain_content.DeptSituation:
      properties:
        lv:
          type: string
        min:
          type: integer
        name:
          type: string
        room:
          type: string
        scenarioId:
          type: string
        tag:
          type: string
        tagCode:
          description: |-
            TagCode is the STATE ("cleared" | "urgent" | "new"); Tag is its label in the
            request's language. They were one Korean field, which forced the client to
            compare against "완료" — so the string could never be translated without
            breaking the comparison that decided which situations count as done.
          type: string
        urgent:
          description: high difficulty
          type: boolean
      type: object
...
    github_com_bingoring_forin_server_internal_domain_home.MentorNote:
      properties:
        id:
          type: string
        npc:
          properties:
            dept:
              type: string
            name:
              type: string
            role:
              type: string
          type: object
        text:
          type: string
      type: object
    github_com_bingoring_forin_server_internal_domain_home.Phrase:
      properties:
        en:
          type: string
        id:
          type: string
        ko:
          type: string
        note:
          type: string
      type: object
    github_com_bingoring_forin_server_internal_domain_home.Shift:
      properties:
        deptLabel:
          type: string
        shift:
          description: DAY | EVENING
          type: string
      type: object
...
    internal_adapters_http.homeColleague:
      properties:
        activeToday:
          type: boolean
        activity:
          type: string
        id:
          type: string
        name:
          type: string
        relation:
          $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_colleague.Relation'
      type: object
    internal_adapters_http.homeResp:
      properties:
        colleagueTotal:
          type: integer
        colleagues:
          items:
            $ref: '#/components/schemas/internal_adapters_http.homeColleague'
          type: array
          uniqueItems: false
        date:
          type: string
        done:
          type: boolean
        level:
          type: integer
        mentorNote:
          $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_home.MentorNote'
        pendingRequests:
          type: integer
        phrase:
          $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_home.Phrase'
        review:
          $ref: '#/components/schemas/internal_adapters_http.homeReview'
        shift:
          $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_home.Shift'
        situationsWaiting:
          type: integer
        streak:
          type: integer
        targetLevel:
          type: string
        todayOne:
          $ref: '#/components/schemas/internal_adapters_http.homeTodayOne'
        unreadCheers:
          type: integer
        week:
          description: |-
            A rolling window ending today (progress.StreakWindowDays long), not a
            calendar week. Kept as `week` on the wire so already-shipped clients keep
            parsing it; the length is what changed, and the client reads .length.
          items:
            type: integer
          type: array
          uniqueItems: false
        xp:
          type: integer
      type: object
    internal_adapters_http.homeReview:
      properties:
        front:
          type: string
        id:
          type: string
      type: object
    internal_adapters_http.homeTodayOne:
      properties:
        chapter:
          type: string
        kind:
          type: string
        scenarioId:
          type: string
        title:
          type: string
      type: object
...
  securitySchemes:
    Bearer:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: forin server API — Go stdlib, hexagonal. Contract is Go-first (swag
    → openapi → openapi-typescript).
  title: forin API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /me/daily-board:
    get:
      responses: null
      security:
      - Bearer: []
      summary: Personalized daily pool — weighted, persisted, resets 00:00 local (?tz=)
      tags:
      - content
  /me/daily-board/topup:
    post:
      responses: null
      security:
      - Bearer: []
      summary: Rewarded-ad top-up of today's daily pool (+N, up to a daily cap)
      tags:
      - content
  /me/home:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapters_http.homeResp'
          description: OK
      security:
      - Bearer: []
      summary: Home tab (one round trip)
      tags:
      - home
...
  /me/route:
    get:
      responses: null
      security:
      - Bearer: []
      summary: Main-route curriculum path (events + unlock states) for the user
      tags:
      - content
  /me/situations:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_content.DeptSituation'
                  type: array
                type: object
          description: OK
      security:
      - Bearer: []
      summary: Department situation cards (?dept=ER) — dept-scoped scenarios
      tags:
      - content
...